refactor/#83 : response 폴더 생성
