+++ api/src/test/java/io/eagle/domain/vacation/service/MarketServiceTest.java
@@ -4,7 +4,7 @@ import io.eagle.common.TestUtil;
 import io.eagle.domain.interest.repository.InterestRepository;
 import io.eagle.domain.picture.repository.PictureRepository;
 import io.eagle.domain.transaction.repository.TransactionRepository;
-import io.eagle.domain.vacation.dto.MarketInfoDto;
+import io.eagle.domain.vacation.dto.response.MarketInfoDto;
 import io.eagle.domain.vacation.repository.VacationRepository;
 import io.eagle.entity.User;
 import io.eagle.entity.Vacation;
